Fix failing AtomicLong test on Linux-x86

+++ unit-tests/shared/src/test/scala/javalib/util/concurrent/atomic/AtomicLongTest.scala
@@ -235,7 +235,7 @@ class AtomicLongTest extends JSR166Test {
    */
   @Test def testFloatValue(): Unit = {
     val ai = new AtomicLong
-    assertEquals(0.0f, ai.floatValue, delta)
+    assertEquals(0.0f, ai.floatValue, delta.toFloat)
     for (x <- VALUES) {
       ai.set(x)
       assertEquals(x.toFloat, ai.floatValue, delta)
